+++ .gitignore
@@ -1,0 +1,31 @@
+tests/.Rhistory
+Python/__pycache__/UTASTAR.cpython-37.pyc
+Python/__pycache__/UTASTAR.cpython-37.pyc
+Python/__pycache__/UTASTAR.cpython-37.pyc
+Python/__pycache__/UTADIS.cpython-37.pyc
+Python/__pycache__/TOPSIS.cpython-37.pyc
+PyTests/__pycache__/TOPSIS.cpython-37.pyc
+Python/__pycache__/TOPSIS.cpython-37.pyc
+Python/__pycache__/UTADIS.cpython-37.pyc
+Python/__pycache__/UTASTAR.cpython-37.pyc
+Python/__pycache__/UTASTAR.cpython-37.pyc
+Python/__pycache__/UTADIS.cpython-37.pyc
+Python/__pycache__/TOPSIS.cpython-37.pyc
+Python/__pycache__/UTASTAR.cpython-37.pyc
+Python/__pycache__/UTADIS.cpython-37.pyc
+Python/__pycache__/TOPSIS.cpython-37.pyc
+Python/__pycache__/UTASTAR.cpython-37.pyc
+Python/__pycache__/UTADIS.cpython-37.pyc
+Python/__pycache__/TOPSIS.cpython-37.pyc
+Python/__pycache__/UTASTAR.cpython-37.pyc
+Python/__pycache__/TOPSIS.cpython-37.pyc
+Python/__pycache__/UTADIS.cpython-37.pyc
+Python/__pycache__/UTASTAR.cpython-37.pyc
+Python/__pycache__/UTADIS.cpython-37.pyc
+Python/__pycache__/TOPSIS.cpython-37.pyc
+Python/__pycache__/UTASTAR.cpython-37.pyc
+Python/__pycache__/UTADIS.cpython-37.pyc
+Python/__pycache__/TOPSIS.cpython-37.pyc
+Pyth/__pycache__/TOPSIS.cpython-37.pyc
+Pyth/__pycache__/UTADIS.cpython-37.pyc
+Pyth/__pycache__/UTASTAR.cpython-37.pyc
