+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "python.pythonPath": "C:\\Users\\Alex_\\AppData\\Local\\Continuum\\anaconda3\\envs\\thesis\\python.exe"
+    "python.pythonPath": "C:\\Users\\Alex_\\AppData\\Local\\Continuum\\anaconda3\\envs\\dip\\python.exe"
 }
